+++ Runtime/Core/ComponentsStorage/LinearStructRegistry/StructComponentsBlittable.cs
@@ -15,8 +15,6 @@ namespace ME.ECS {
 
         [ME.ECS.Serializer.SerializeField]
         internal NativeBufferArraySliced<Component<TComponent>> components;
-        [ME.ECS.Serializer.SerializeField]
-        private long maxVersion;
 
         public override UnsafeData CreateObjectUnsafe(in Entity entity) {
             
@@ -85,7 +83,6 @@ namespace ME.ECS {
                 var v = (long)this.world.GetCurrentTick();
                 ref var data = ref this.components[entity.id];
                 data.version = v;
-                this.maxVersion = (v > this.maxVersion ? v : this.maxVersion);
             }
 
         }
@@ -97,7 +94,6 @@ namespace ME.ECS {
 
             if (AllComponentTypes<TComponent>.isVersioned == true) {
                 bucket.version = this.world.GetCurrentTick();
-                this.maxVersion = (bucket.version > this.maxVersion ? bucket.version : this.maxVersion);
             }
 
         }
@@ -123,7 +119,6 @@ namespace ME.ECS {
         public override void OnRecycle() {
 
             this.components = this.components.Dispose();
-            this.maxVersion = default;
             base.OnRecycle();
             
         }
@@ -141,13 +136,7 @@ namespace ME.ECS {
 
         public override IComponentBase GetObject(Entity entity) {
 
-            #if WORLD_EXCEPTIONS
-            if (entity.IsAlive() == false) {
-                
-                EmptyEntityException.Throw(entity);
-                
-            }
-            #endif
+            E.IS_ALIVE(in entity);
 
             var index = entity.id;
             ref var bucket = ref this.components[index];
@@ -163,13 +152,7 @@ namespace ME.ECS {
 
         public override bool SetObject(in Entity entity, UnsafeData buffer, StorageType storageType) {
             
-            #if WORLD_EXCEPTIONS
-            if (entity.IsAlive() == false) {
-                
-                EmptyEntityException.Throw(entity);
-                
-            }
-            #endif
+            E.IS_ALIVE(in entity);
 
             return DataBlittableBufferUtils.PushSet_INTERNAL(this.world, in entity, this, buffer.Read<TComponent>(), storageType);
 
@@ -177,13 +160,7 @@ namespace ME.ECS {
 
         public override bool SetObject(in Entity entity, IComponentBase data, StorageType storageType) {
 
-            #if WORLD_EXCEPTIONS
-            if (entity.IsAlive() == false) {
-                
-                EmptyEntityException.Throw(entity);
-                
-            }
-            #endif
+            E.IS_ALIVE(in entity);
 
             return DataBlittableBufferUtils.PushSet_INTERNAL(this.world, in entity, this, (TComponent)data, storageType);
             
@@ -262,7 +239,6 @@ namespace ME.ECS {
             base.CopyFrom(other);
             var _other = (StructComponentsBlittable<TComponent>)other;
             NativeArrayUtils.Copy(in _other.components, ref this.components);
-            this.maxVersion = _other.maxVersion;
 
         }