+++ Runtime/Core/Entities/EntitiesGroup.cs
@@ -435,7 +435,7 @@ namespace ME.ECS {
             var data = new Component<TComponent>() {
                 data = (TComponent)component,
                 state = 1,
-                version = ++this.maxVersion + 1,
+                version = 1,
             };
             var componentIndex = ComponentTypes<TComponent>.typeId;
             ref var archetypes = ref this.world.currentState.storage.archetypes;
@@ -462,7 +462,7 @@ namespace ME.ECS {
             var data = new Component<TComponent>() {
                 data = component,
                 state = 1,
-                version = ++this.maxVersion + 1,
+                version = 1,
             };
             var componentIndex = ComponentTypes<TComponent>.typeId;
             ref var archetypes = ref this.world.currentState.storage.archetypes;
@@ -523,14 +523,8 @@ namespace ME.ECS {
         /// <returns></returns>
         public EntitiesGroup AddEntities(int count, Unity.Collections.Allocator allocator, bool copyMode) {
             
-            #if WORLD_STATE_CHECK
-            if (this.HasStep(WorldStep.LogicTick) == false && this.HasResetState() == true) {
-
-                OutOfStateException.ThrowWorldStateCheck();
-                
-            }
-            #endif
-
+            E.IS_LOGIC_STEP(this);
+            
             var group = new EntitiesGroup();
             if (count <= 0) return group;