+++ Runtime/Core/Utils/RandomUtils.cs
@@ -16,19 +16,6 @@ namespace ME.ECS {
 
     public static class RandomUtils {
 
-        /// <summary> Typical usage: RandomUtils.ThreadCheck(currentWorld.worldThread); </summary>
-        #if INLINE_METHODS
-        [System.Runtime.CompilerServices.MethodImplAttribute(System.Runtime.CompilerServices.MethodImplOptions.AggressiveInlining)]
-        #endif
-        [System.Diagnostics.Conditional("WORLD_THREAD_CHECK")]
-        public static void ThreadCheck(World world) {
-            #if WORLD_THREAD_CHECK
-            if (world.worldThread != System.Threading.Thread.CurrentThread) {
-                WrongThreadException.Throw("Random", "this could cause sync problems");
-            }
-            #endif
-        }
-
         #if INLINE_METHODS
         [System.Runtime.CompilerServices.MethodImplAttribute(System.Runtime.CompilerServices.MethodImplOptions.AggressiveInlining)]
         #endif